Removed profiling from main system and created profile runner - as yes untested
