+++ lib/directoryWatcher.js
@@ -20,10 +20,9 @@ var chokidar = require('chokidar')
 module.exports = function () {
 
   var start = function (container, changedCb, cb) {
-    // todo expose follow sym links and depth as options in the config file default to the below
+    // var watcherArgs = {persistent: true, ignoreInitial: true, followSymlinks: false, depth: 4, usePolling: true}
     var watcherArgs = {persistent: true, ignoreInitial: true, followSymlinks: false, depth: 4}
 
-    // do not monitor docker conatiners
     if (container.type === 'docker' || container.type === 'container') {
       return cb(null)
     }
