+++ lib/system.js
@@ -1,3 +1,4 @@
+/* jshint latedef: false */
 /*
  * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESSED OR IMPLIED
  * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
@@ -14,13 +15,9 @@
 
 'use strict'
 
-/* jshint latedef: false */
-
 var _ = require('lodash')
 var async = require('async')
 var chalk = require('chalk')
-var path = require('path')
-var spawn = require('child_process').spawn
 var runner = require('./runner')()
 var directoryWatcher = require('./directoryWatcher')()
 var util = require('./util')()
@@ -79,13 +76,13 @@ module.exports = function () {
               if (err) { return console.log(err) }
 
               processes[child.pid] = {container: proc.container,
-                                      identifier: proc.container.name,
-                                      running: true,
-                                      exitCode: null,
-                                      colour: util.selectColour(),
-                                      child: child,
-                                      monitor: proc.monitor,
-                                      tail: proc.tail}
+                identifier: proc.container.name,
+                running: true,
+                exitCode: null,
+                colour: util.selectColour(),
+                child: child,
+                monitor: proc.monitor,
+                tail: proc.tail}
               util.streamOutput(processes[child.pid], _system.global.log_path)
               directoryWatcher.start(proc.container, directoryChangeCb(child.pid), function (watcher) {
                 processes[child.pid].watcher = watcher
@@ -102,13 +99,13 @@ module.exports = function () {
             if (err) { return console.log(err) }
 
             processes[child.pid] = {container: proc.container,
-                                    identifier: proc.container.name,
-                                    running: true,
-                                    exitCode: null,
-                                    colour: util.selectColour(),
-                                    child: child,
-                                    monitor: proc.monitor,
-                                    tail: proc.tail}
+              identifier: proc.container.name,
+              running: true,
+              exitCode: null,
+              colour: util.selectColour(),
+              child: child,
+              monitor: proc.monitor,
+              tail: proc.tail}
             util.streamOutput(processes[child.pid], _system.global.log_path)
             directoryWatcher.start(proc.container, directoryChangeCb(child.pid), function (watcher) {
               processes[child.pid].watcher = watcher
@@ -137,16 +134,16 @@ module.exports = function () {
 
             if (child) {
               processes[child.pid] = {container: container,
-                                      identifier: container.name,
-                                      running: true,
-                                      exitCode: null,
-                                      colour: procColour,
-                                      child: child,
-                                      startTime: Date.now(),
-                                      monitor: container.monitor,
-                                      tail: container.tail}
-              util.streamOutput(processes[child.pid], _system.global.log_path)
+                identifier: container.name,
+                running: true,
+                exitCode: null,
+                colour: procColour,
+                child: child,
+                startTime: Date.now(),
+                monitor: container.monitor,
+                tail: container.tail}
 
+              util.streamOutput(processes[child.pid], _system.global.log_path)
               directoryWatcher.start(container, directoryChangeCb(child.pid), function (watcher) {
                 processes[child.pid].watcher = watcher
                 checkDelay(next)
@@ -211,51 +208,6 @@ module.exports = function () {
 
 
 
-  var profile = function (system, args, cb) {
-    if (process.platform === 'win32') {
-      console.log('Sorry! Profiling isn\'t supported on Windows')
-      return cb()
-    }
-    var name = args.process
-    var container = _.find(system.topology.containers, function (c) { return c.name === name })
-    if (container && util.isNodeProcess(container)) {
-
-      spawn('sudo', ['-n', 'true'])
-        .on('exit', function (code) {
-          if (code !== 0) {
-            console.log('Profiling requires sudo access:')
-          }
-
-          spawn('sudo', ['true'])
-            .on('exit', function (code) {
-              if (code) {
-                console.log('The profiler cannot be executed without sudo access.')
-                return cb()
-              }
-              // todo reliably find 0x bin (maybe 0x does require('0x/where'))
-              var flags = Object.keys(args.options).map(function (k) {
-                return '--' + k + '=' + args.options[k]
-              }).join(' ')
-
-              var cmd = 'node ' + path.join(__dirname, '..', 'run-0x') + ' -q ' + flags + ' node '
-              container._cachedRun = container.run
-              container.run = container.run.replace(/node/g, cmd)
-              container._cachedMonitorStatus = container.monitor
-              container.monitor = false
-              container.profiling = true
-              start(system, name, 1, function () {
-                cb()
-              })
-            })
-        })
-    } else {
-      console.log('Only node processes can be profiled!')
-      cb()
-    }
-  }
-
-
-
   var stop = function (system, name, count, cb) {
 
     var waitStop = function (pid, next) {
@@ -276,25 +228,13 @@ module.exports = function () {
         stoppingProcess = process.identifier
         directoryWatcher.stop(process.watcher)
 
-        if (process.container.profiling) {
-          runner.kill(process.container, process.child.pid, 'SIGINT', function () {
-            container.profiling = false
-            container.monitor = container._cachedMonitorStatus
-            container.run = container._cachedRun
-            delete container._cachedRun
-            delete container._cachedMonitorStatus
-            waitStop(process.child.pid, next)
-          })
-        } else {
-          runner.stop(process.container, process.child.pid, function () {
-            if (container.debug) {
-              container.run = container.run.replace(/node-debug -d [0-9]+ -p [0-9]+/g, 'node')
-              container.debug = false
-            }
-            waitStop(process.child.pid, next)
-          })
-        }
-
+        runner.stop(process.container, process.child.pid, function () {
+          if (container.debug) {
+            container.run = container.run.replace(/node-debug -d [0-9]+ -p [0-9]+/g, 'node')
+            container.debug = false
+          }
+          waitStop(process.child.pid, next)
+        })
       } else {
         next('process not found')
       }
@@ -422,19 +362,6 @@ module.exports = function () {
 
 
 
-  var heap = function (system, name) {
-    var proc = util.findProcess(processes, name)
-    if (!proc || !proc.child || typeof proc.child.send !== 'function') {
-      console.log('connected child process not found')
-      return
-    }
-
-    console.log('creating heap snapshot: ' + name)
-    proc.child.send({ command: 'heapdump-create' })
-  }
-
-
-
   var sendMessage = function (system, name, data) {
     var proc = util.findProcess(processes, name)
     if (!proc || !proc.child || typeof proc.child.send !== 'function') {
@@ -448,47 +375,10 @@ module.exports = function () {
 
 
 
-  var messageHandler = function (message) {
-    if (message && typeof message === 'object') {
-      message = JSON.stringify(message, null, '  ')
-    }
-
-    console.log(message)
-  }
-
-
-
-  var startReport = function (system, name) {
-    var proc = util.findProcess(processes, name)
-    if (!proc || !proc.child) {
-      console.log('child process not found')
-      return
-    }
-
-    console.log('Reporting messages for: ' + name)
-    proc.child.on('message', messageHandler)
-  }
-
-
-
-  var stopReport = function (system, name) {
-    var proc = util.findProcess(processes, name)
-    if (!proc || !proc.child) {
-      console.log('child process not found')
-      return
-    }
-
-    console.log('Stopping messages for: ' + name)
-    proc.child.removeListener('message', messageHandler)
-  }
-
-
-
   return {
     startAll: startAll,
     start: start,
     debug: debug,
-    profile: profile,
     stopAll: stopAll,
     stop: stop,
 
@@ -505,9 +395,6 @@ module.exports = function () {
     processes: procs,
     reset: reset,
 
-    heap: heap,
-    sendMessage: sendMessage,
-    startReport: startReport,
-    stopReport: stopReport
+    sendMessage: sendMessage
   }
 }
